+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add metadata for userscript auto-updates [#27]
 - Add support for javdb
 - Improve responsiveness for sites with many entries by queueing batch queries [#28]
+- Fix data18 performer overview page
 
 ## [0.9.3]